refactor: test feature error recovery in associated header file
#1
