+++ test/datasets/MinimalV3AnsFrame.h
@@ -1,5 +1,5 @@
-#ifndef TEST_DATASET_MINIMAL_V3_ANS_FRAME_HPP
-#define TEST_DATASET_MINIMAL_V3_ANS_FRAME_HPP
+#ifndef TEST_DATASET_MINIMAL_V3_ANS_FRAME_H
+#define TEST_DATASET_MINIMAL_V3_ANS_FRAME_H
 #include <array>
 // clang-format off
 constexpr std::array<uint8_t,0x0d> MINIMAL_V3_ANS_FRAME {
@@ -10,4 +10,4 @@ constexpr std::array<uint8_t,0x0d> MINIMAL_V3_ANS_FRAME {
     0x00, 0x00, 0x03, 0xd7, /* checksum */
 };
 // clang-format on
-#endif // TEST_DATASET_MINIMAL_V3_ANS_FRAME_HPP
+#endif // TEST_DATASET_MINIMAL_V3_ANS_FRAME_H
