+++ test/datasets/errors/BadAnswerProtocolVersion.h
@@ -1,5 +1,5 @@
-#ifndef TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_HPP
-#define TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_HPP
+#ifndef TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_H
+#define TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_H
 #include <array>
 // clang-format off
 constexpr std::array<uint8_t,0x0d> BAD_ANSWER_PROTOCOL_VERSION {
@@ -10,4 +10,4 @@ constexpr std::array<uint8_t,0x0d> BAD_ANSWER_PROTOCOL_VERSION {
     0x00, 0x00, 0x3, 0xd4, /* checksum */
 };
 // clang-format on
-#endif //TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_HPP
+#endif //TEST_DATASET_BAD_ANSWER_PROTOCOL_VERSION_H
